+++ src/UserRunAction.cc
@@ -12,8 +12,8 @@ UserRunAction::~UserRunAction()
 
 void UserRunAction::EndOfCurrentStepOfRun()
 {
-	/* int numtracks = RunManager::getInstance()->GetTrackPropagation()->SizeofTrackList();
-    for (int i = 0; i < numtracks; ++i)
+	/* size_t numtracks = RunManager::getInstance()->GetTrackPropagation()->SizeofTrackList();
+    for (size_t i = 0; i < numtracks; ++i)
     {
         Track *track = RunManager::getInstance()->GetTrackPropagation()->GetTrack(i);
 		if((fabs(track->GetPostStepPoint().energy) < 5.0 * unit_eV) && (track->GetParticleName() == "h")) track->SetShouldBeKilled(true);
@@ -69,8 +69,8 @@ void UserRunAction::EndOfRun()
 	  tout->SetBranchAddress("ctime",&ctime);
 	}
 
-	int numtracks = RunManager::getInstance()->GetTrackPropagation()->SizeofTrackList();
-	for (int i = 0; i < numtracks; ++i)
+	size_t numtracks = RunManager::getInstance()->GetTrackPropagation()->SizeofTrackList();
+	for (size_t i = 0; i < numtracks; ++i)
 	{
 	    Track *track = RunManager::getInstance()->GetTrackPropagation()->GetTrack(i);
